Exceptions moved to database package. Exceptions will be thrown by database and caught in and handled in Controller package in future iteration.
